move get_random_array to hub.tests.common
