+++ hub/core/tests/common.py
@@ -1,6 +1,6 @@
 import pytest
 
-from typing import Dict, List, Tuple
+from typing import Dict, List
 
 import numpy as np
