+++ hub/tests/common.py
@@ -1,6 +1,8 @@
 import pytest
+import numpy as np
 import os
 from uuid import uuid1
+from typing import Tuple
 
 from hub.constants import B, KB, MB
 
@@ -44,3 +46,39 @@ def current_test_name() -> str:
     test_name = full_name.split("::")[1]
     output = os.path.join(test_file, test_name)
     return output
+
+
+def get_random_array(shape: Tuple[int], dtype: str) -> np.ndarray:
+    dtype = dtype.lower()
+
+    if "int" in dtype:
+        low = np.iinfo(dtype).min
+        high = np.iinfo(dtype).max
+        return np.random.randint(low=low, high=high, size=shape, dtype=dtype)
+
+    if "float" in dtype:
+        # get float16 because np.random.uniform doesn't support the `dtype` argument.
+        low = np.finfo("float16").min
+        high = np.finfo("float16").max
+        return np.random.uniform(low=low, high=high, size=shape).astype(dtype)
+
+    if "bool" in dtype:
+        a = np.random.uniform(size=shape)
+        return a > 0.5
+
+    raise ValueError("Dtype %s not supported." % dtype)
+
+
+@parametrize_dtypes
+@pytest.mark.parametrize(
+    SHAPE_PARAM,
+    (
+        (100, 100),
+        (1,),
+        (1, 1, 1, 1, 1),
+    ),
+)
+def test_get_random_array(shape: Tuple[int], dtype: str):
+    array = get_random_array(shape, dtype)
+    assert array.shape == shape
+    assert array.dtype == dtype
